NRFFOSDK-11439: Add p_context to nrfx_spis API
The parameter set on initialization of the driver is then passed
to the event handler. This aligns the SPIS driver with other
multi-instance ones (TWI, UART, SPI master). For the same reason
(and to improve efficiency) the event is now passed to the handler
by a pointer and it is allowed for zero-length buffers to pass NULL
instead of a valid pointer.