NRFFOSDK-NONE: Fix nrfx_adc_uninit()
After _all_channels_disable() was added, _uninit() was switched
to call it instead of just setting the list head pointer to NULL.
This turned out to be not a good idea, since _all_channels_disable()
cannot be called when the ADC is busy, while _uninit() is supposed
(as stated in the documentation) to deal with ongoing conversions
(stop them all). This has been corrected.
Moreover, the disabling of all channels was moved after the disabling
of the interrupt. Otherwise, the NULL pointer could be dereferrenced.

Documentation of channel enabling and disabling functions was enhanced
as well.