NRFFOSDK-NONE: Extend SAADC HAL with a PSELP-only setter
The need for modifying only the positive input for a given channel appeared
in the ADC driver for Zephyr (the setting of the positive input enables or
disables the whole channel, at least according to the nRF52832 documentation,
for nRF52840 this section has been rephrased and no longer states this fact).
A similar function for the negative input pin seemed to be useless, thus it
has not been added.