NCSDK-515: Add static assert to check NUMBER_OF_PINS symbol
It seems that driver works correctly as long as NUMBER_OF_PINS value is lesser
or equal to 64, because 6 bits are used to encode pin number. As for now, static
assert has been added in order to raise an error when attempting to enable driver
on chip with more than 64 pins (which is not the issue with current hardware).