NRFFOSDK-12683: prevent race condition
Secure against the race condition that occurs when an IN transfer is interrupted by an OUT
transaction, which in turn is interrupted by a process with higher priority. If the IN events
ENDEPIN and EPDATA arrive during that high priority process, the OUT handler might call
usbd_ep_data_handler without calling nrf_usbd_epin_dma_handler
(or nrf_usbd_ep0in_dma_handler) for the IN transaction.