nrfx_usbd: Add an alternative way of ending the DMA transfer loop
The waiting loop in `usbd_dmareq_process` is intended to prevent
accesses to any USBD registers while a DMA transfer is in progress.
Normally it should finish when the END event for the transfer is
generated. Apparently, there are cases when this does not happen
and this loop waits indefinitely. This patch provides an alternative
way of exiting this loop, when the USBRESET event occurs.