NCSDK-2857: Enhance nrfx_gpiote with pin task and event getters
Currently the nrfx_gpiote driver provides only functions that return
addresses of the GPIOTE events or tasks associated with specified pins.
These addresses are sufficient for configuring PPI channels, but in
case of DPPI the more desirable things to obtain are the enumerated
values for these events and tasks (which are in fact their offsets in
the GPIOTE peripheral address space), as the `nrf_gpiote_publish_set`
and `nrf_gpiote_subscribe_set` functions require such parameters.
This commits adds functions that return the enumerated values.