NCSDK-4013: Improve FTPAN-86 workaround handling in nrfx_saadc
FTPAN-86 states that spurious samples can be written into RAM during
calibration. Workaround for nRF53 & nRF91 is to trigger STOP task
before each reconfiguration of the channels. It appears that if burst
or oversampling were enabled before, then this workaround does not
work and different one has to be applied.
nRF51 & nRF52 Series are not affected as power cycle is enough.