NCSDK-NONE: Fix nrfx_pwm_stop
Deactivate shortcuts before triggering the STOP task, to prevent
the PWM from being immediately started again if the LOOPSDONE event
occurs at the same cycle as the STOP task is triggered. And trigger
the STOP task always, even if the PWM appears to be already stopped.
It was observed by a customer that in non-handler mode, very rarely
but still, the PWM was not stopped when triggering of the STOP task
was performed conditionally.