NCSDK-4203: Fix TWIM driver to handle mixing of standard and hold transfers
Driver failed when hold (repeated) transfer was setup but never
started, followed by setting standard transfer (triggered). That
happened in case workaround for anomaly 109 was applied.

Rootcause was in the workaround which artifiacially set frequency
to zero before transfer was started. When transfer was never
triggered, frequency remained set to zero. New standard transfer
setup did not restore frequency and TWIM was not working.

Additionally, driver was applying workaround for Anomaly 109 only
for the first transfer because a trick with frequency set to 0
was applied only in transfer setup. Added same workaround to
handling of stopped event to ensure that following transfers
have workaround applied.