NRFFOSDK-10704: nrfx_nfct: change the order of tx events handled in IRQ handler
The TXFRAMESTART event should be handled before the TXFRAMEEND event in
the NFCT IRQ Handler. In case of very short TX packets, both event flags
are set nearly simultaneously and are handled by the IRQ handler in one
go. In this scenario, the TXFRAMEEND callback is called before the
TXFRAMESTART callback.