nrfx_ncft: Add function to tx specific bit count
Out of the box, nrfx_nfct only had the functionality to transmit whole
bytes. Things in many forms of NFC communication depend on specific bit
counts (like ACK/NAKs for particular protocols.) This adds a function
with a signature similar to nrfx_nfct_tx, aptly named nrfx_nfct_bits_tx.
It uses the data_size member of the nrfx_nfct_data_desc_t struct as
the number of bits to send rather than bytes. As such, it also
calculates the byte length of the buffer to provide to the MAXLEN
EasyDMA register based on the number of bits, taking into account bits
that don't fit evenly into the last whole byte.

Signed-off-by: Andreas Pettersson <andreaspettersson95@gmail.com>