NCSDK-14734: Increase SPU.FEATURE.GPIO to max superset size
GPIO feature is now a superset of all possible ports in Haltium
products. UICR GPIO allocation in nRF92 allows up to P12, which is out
of bounds in MDK 8.45.0 even though the registers exist in hardware.

This change allows SDFW to still use the HAL to allocate all possible
ports in SPU instances with this feature.

See reference task for related PR to have this officially brought into
the MDK.

Signed-off-by: Stephen Stauts <stephen.stauts@nordicsemi.no>