NRFX-NONE: Add missing `|| defined(__NRFX_DOXYGEN__)` in driver headers
Doxygen does not process definitions of symbols from files included
by the processed file, so symbols like `NRF_CLOCK_HAS_HFCLK192M`
(that are defined as 1 when `__NRFX_DOXYGEN__` is defined) can be
used for conditional processing of Doxygen stuff only in the file
where they are defined (in this case, only in nrf_clock.h, and not
in e.g. nrfx_clock.h).

Add missing `defined(__NRFX_DOXYGEN__)` conditions in driver headers
that define some functions or structure members basing on symbols
defined in HAL or MDK headers.