NRFX-NONE: Cast return values to uint8_t/uint16_t
In some cases, registers values are returned. Most registers are uint32_t.
This PR casts the return values to uint8_t to avoid compiler warnings.

Addressed also a case where arithmetic operations resulted in a uint32_t, but the return value is of type uint16_t.

The casts are added to avoid compiler warnings.

Signed-off-by: Johan Stridkvist <johan.stridkvist@nordicsemi.no>