NRFX-NONE: nrf_hsfll: Flip MIRROR register logic
Previously mirror lock logic was flipped but it was a
misinterpretation of the documentation combained with specific
behavior when closed loop mode is enabled (with closed loop
applying FICR with mirror locked was not giving expected
frequency).

After discussion with Tarpan and SAG it has been clarified
that when MIRROR lock is disabled then registers are applied
immediately one by one vs batch appling when lock was enabled
during setting and disabled after all registers have been
updated.