NRFX-2618: nrfx_common: Add always_inline to RISCV __STATIC_INLINE
In general, we expect that all hal and haly functions will
be inlined because they are often called with fixed parameter
and compiler can reduce the code during the compliation.
However, it's been seen that riscv gcc complier does not
inline them if there are multiple calls to the same hal(y)
function. As a result HAL(Y) function is not optimized and
performance is degraded and code is bigger (uarte driver
as an example).