NRFX-NONE: nrfx_uarte: Fix case when RX is disabled in callback
Driver was not handling case when RX was aborted from RX_RDY event.
That happened because check for abort in ENDRX event handler was
done before user handler was called. If user aborted RX from that
context then ENDRX handler was not aware of that. Added another
check after user handler.