NRFX-NONE: nrfx_grtc: Limit workaround only for RISCV cores
There is a workaround in the irq_handler that disables the event
for channel 0. According to the comments in HM-15402 it applies
only to RISCV. Because of this workaround, it is possible to
trigger a scenario where GRTC stops working. This may happen if
GRTC is configured from higher interrupt context than GRTC
interrupt. If new CC is set for channel 0 after GRTC event
triggered but before event is disabled then new value will no
longer trigger an interrupt (because event just got disabled).