NRFX-4664: nrfx_uarte: Rework ENDRX handling
Previously, ENDRX handling was attempting to detect case when previous
ENDRX has not handled on time. If that was detected then events were
reported as if 2 ENDRX occurred (even though only one was captured).
Detection was based on RXSTARTED interrupt flag being set in the same
interrupt handling execution as ENDRX assuming that RXSTARTED indicates
that there was ENDRX_STARTRX short. However, it is not always true,
RXSTARTED can be set because interrupt handler was unable to execute
between starting current transfer and end of that transfer. Because of
that case, detection is removed and driver do not assume anything.

Additional optimization in ENDRX handling lead to >90 bytes of code
saving.