NRFX-4926: Allow on-the-fly I2S buffer size update
It is mandatory to allow on-the-fly buffer size updates in order to
fully adhere to Zephyr I2S API. Prominent use-case is Asynchronous USB
Audio where there is a need to keep I2S FRAMESTART as close as possible
to USBD SOF while the audio data is actively streamed. To slighly delay
or advance I2S FRAMESTART relative to USBD SOF the host can, in response
to asynchronous endpoint feedback, supply one more or one less sample in
data packet.

Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>