NRFX-5062: nrfx_uarte: Fix corner case for rx_new_data_check
It is possible that there is some data that is stored during flushing.
That data is copied into the beginning of the next buffer (if
NRFX_UARTE_RX_ENABLE_KEEP_FIFO_CONTENT was set). If data was copied
then next nrfx_uarte_rx_new_data_check() shall return true even when
RXDRDY HW event is not set. Without this new data is kept in the
buffer unnoticed.