NRFX-5090: nrfx_uarte: Allow for using TX_LINK without PPI
nrfx_uarte API had TX_LINK flag which was enabling linking TX
transfers to get the best utilization of the UART bandwidth.
It was required to use (D)PPI with that. However, tests have shown
that even if PPI is not used but next TX transfer is started
immediately from ENDTX event handler there is a significant gain
compared to standard approach where next transfer is started
from TX_DONE event. For example, on nrf52 test transfer at 1M
was 25% shorter with PPIed linking and 20% shorter with SW
linking.