DRGN-20749: nrfx_uarte: Fix race condition in RX path
There are a race condition when RXSTARTED event arrived after check
for that event in the interrupt handler but when ENDRX is already
detected. In that case, driver was assuming that new rx buffer was
set too late (after ENDRX) and was manually triggering the transfer.
However, transfer was correctly triggered so it was restarted which
resulted in some bytes being dropped.