NRFX-NONE: nrfx_uarte: Improve handling notifying about flushed bytes
There was a notification about RX_BYTE called directly from
rx_flushed_handler if RXDRDY interrupt was enabled. This would never
happen because all interrupts sources are disabled in that function.
As a consequence driver was not notifying about pending byte after
enabling the receiver. Moving that notification to
nrfx_uarte_rx_enable where interrupts are not masked.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>