NRFX-5691: Fix nrfx_twim err code for 1 MHz pins when GPIO cfg skipped
On nRF5340 it is possible to use 1 MHz clock frequency for TWIM,
provided that dedicated pins are used. If dedicated pins are not used
driver should return an error. However, if GPIO and PSEL config
is to be skipped no error code should be returned, as driver
configuration structure might not contain valid pin numbers.
On the occasion fix missing pin config in nrfx_twim_reconfigure().