NRFX-NONE: Get rid of redundant GRTC event clears
It is allowed to set the SYSCOUNTER CC to the value from the past. Thus
the event should not be cleared after the `cc_set` or `cc_add_set` call
because such event will be missed.
Also there is no need to enable the compare event then because writing
to CCH or CCADD register enables it automatically.

Signed-off-by: Adam Kondraciuk <adam.kondraciuk@nordicsemi.no>