NRFX-5863: nrfx_uarte: Fix blocking abort case
Previous fix with checking if peripheral is enabled and if not returning
amount in nrfx_uarte_rx_ready was wrong as it would always return last
amount register value when peripheral was disabled. Reworked
nrfx_uarte_rx_ready to return NRFX_ERROR_ALREADY when peripheral is
disabled and adjuster nrfx_uarte_rx function to handle it correctly.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>