NRFX-NONE: nrfx_uarte: Rearrange static function location
Move static function up in the code so it can be used in functions
following the new location.

Additionally, improve RX cache case so that internal buffers
are updated before user handler is executed so that in
the context of the user handler buffers are already switched
(next is written to current and next is null) and new buffer
can be provided.