NRFX-NONE: nrfx_uarte: Fix RX disabling from flushing
If RX was disabled from flushing (because there is more bytes from
FIFO then input buffer) then another enabling might have been called
from the context of RX disable event. In that case, driver was faulting
because it was in the middle of enabling when another enabling was
happening.

Additionally, when RX flushed data is enough to fill the new buffer
then RX DONE event shall be called to the user. However, wrong
function was called, the one that did not take into account handling
of RX cache (bounce buffers).