NRFX-NONE: nrfx_uarte: Improve RX flushing
When user is aborting RX then it means that there is no intrest
in any remaining data. In that case RX FIFO content shall be
dropped and not prepended to the next RX buffer provided when
receiver is reenabled.

Minor tweak in workaround in RX FIFO flushing. If amount value
is bigger than the RX FIFO size that indicates that FIFO was
empty.