NRFX-NONE: nrfx_uarte: Fix case when RX is disabled prematurely
It is possible that data from RX FIFO flushing is pending and new
on next RX enabling first provided buffer is equal or shorter compared
to the pending RX data. In that case, user buffer will be filled and
data fill be reported immediately and RX will be disabled. Driver
was not ready for that and was expecting to have non-null RX buffer
during RX enabling. Added additional check which returns from RX
enabling function with success if RX is disabled.