NRFX-6308: Switch to AUTOEN GRTC read manner
Change the manner of reading the SYSCOUNTER. AUTOEN option ensures that
the SYSCOUNTER is always active whenever any of CPUs is not sleeping.
When all CPUs are in any of sleep states, the SYSCOUNTER is also switched
to sleep state by hardware. This option should not be disabled until the
user provides the correct SYSCOUNTER reading procedure as described in PS.

Signed-off-by: Adam Kondraciuk <adam.kondraciuk@nordicsemi.no>