NRFX-6554: Add API to nrfx_gppi to setup DPPI interconnect
Added a function that allows the user to setup an interconnect between
two DPPI instances without connecting them to the final endpoints.

The new API re-uses most of the logic from the
nrfx_gppi_channel_endpoints_setup API function, except that it doesn't
connect the DPPIs to the final endpoints and does not take the ownership
over the DPPI channels supplied by the user.

This allows the use-cases, such as connecting statically allocated DPPI
channels to a peripheral in a different power domain, such as connecting
a statically allocated DPPI in radio domain with GPIOTE in the
peripheral domain.