NRFX-4686: Extend DPPI channels ownership
Added owned channels mask. This means that given core can configure and use
those channels inside power domain (or APB that belongs to this domain) which
used DPPIC peripheral belongs to. If an external connection is required
(to other power domain/APB) also publish or subscribe channels mask must
be defined as a subset of owned channels mask.

Signed-off-by: Adam Kondraciuk <adam.kondraciuk@nordicsemi.no>