NRFX-7004: nrfx_clock: Use INTPEND in the interrupt handler
nRF54L got INTPEND register in the clock peripheral. Register can
be used to handle pending interrupts in more efficient way because
only pending events are handled and there is no need to check
if every support EVENT is set. Using this approach is faster and
takes less code (~140 bytes).

Reworked clock interrupt handler to use INTPEND-like approach also
for legacy case. In that case INTEN register is read and then for
every bit that is set in INTEN event is checked. It mimics the
behavior of INTPEND and allows to keep the common implementation.

Part of the rework was to arrange clock events so that they match
bit position for interrupt sources in INTEN register.