NRFX-NONE: nrf_power.h Strict C compliance warning waive
Enumerate values need to fit in the range of int if we are to be
strictly standad compliant.
But for this emum, its highest value is generated from a bitmask
of the 31st bit, which, as a possitive number is out the int range.

This can cause warnings for compilers which are run in a pedantic
enough mode. GCC is one of this, producing a warning like:
"warning: ISO C restricts enumerator values to range of ‘int’"
when run with "-Wpedantic"
Let's disable these warnings for these enums definitions
when building for GCC.
Apart from this warning, GCC seems to properly handle this case.

Note that as this definitions are in headers, the issue is triggered
when a user tries to compile their code with the corresponding warnings
enabled.

The issue was introduced in
ae3e79a0ae38a8f81ed78dd599095b35e0924fd7
Before, the definition of NRF_POWER_RAMPOWER_S15RETENTION_MASK had
the constant cast to (int) to avoid the warning.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>