NRFX-7697: Remove MDK fixup for nRF54LM20A MINTTHRESH
MINTTRESH fixup was added to deal with inconsistencies
related to value to be written to disable threshold
and the read-back value from the register.
Discussion in DLT-4082 has not concluded in a way
the fixup assumed and MDK 8.71.1 still preserves old
behavior. Therefore remove the fixup for now
to align the nrfx with the existing MINTTHRESH approach.
This will unify the testing results between public
and internal CI.